agents/requirements.txt
anthropic
httpx
requests
python-dotenv
pyyaml
